refactor(ui): share date group utility
